requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
sqlalchemy==2.0.30
pysqlite3-binary==0.5.3

# Vector Database
chromadb==0.5.3

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Rate Limiting
slowapi==0.1.9

# HTTP Client (WeChat API, Qwen API)
httpx==0.27.0

# Environment Variables
python-dotenv==1.0.1

# Async support
aiofiles==23.2.1

# Scheduling (Phase 5 extension)
apscheduler==3.10.4

# Validation
pydantic==2.7.1
pydantic-settings==2.3.1

# Baidu ASR
baidu-aip==4.16.13
